Add OBDotNet
OBDotNet is the compiled assembly for the .NET version of Open Babel
(http://openbabel.org/wiki/Main_Page)

I can use this to generate a molecule's 3D coordinates from a SMILES string.

I've just taken the OBDotNet assembly (and its dependencies) from the windows
installation of the OpenBabelGUI. This assembly was compiled for x86 so the project
has to specifically target x86.

I've also added a post-build step to copy the OBDotNet dependencies into the
output folder.