Remove some unused code
With the introduction of the MoleculeBuilder and the use of OBDotNet
there's some code that I don't need now...

The prepared mol files and json file aren't needed now as I just call
into the api with a SMILES string.  (Which also means I don't need the
MolFileReader & LineReader.)

I've also deleted the tests project because it was only testing reading
from mol files.