Use the obabel command line exe for generating the 3D molecule from SMILES
I couldn't get the OBDotNet wrapper to work properly without the Open Babel GUI
installed on the machine. (And I don't want to have to rely on this dependency.)

I may come back and have another go at this, because it's much more convenient
to work with the OBDotNet wrapper (not least, because of having to fiddle with
the current directory in MoleculeBuilder.FromSmiles) but for now, this works
without any external dependencies.

The Mol format is fairly easy to read so I set the exe to output in that format.
I've re-introduced the LineReader and MolReader to read the output.