Update to the latest three.js
I need to do some more involved object manipulation in three.js to handle
multiple bond orders. (It currently just draws every bond as if it was a single
bond.)

While investigating how to do this, I noticed some differences between the
three.js documentation and the code I was using so to avoid the confusion I'm
updating to the latest.

(Which requires an updated TrackballControls as well - from
https://github.com/mrdoob/three.js/tree/master/examples/js/controls)