Include bond order in the bond data
In preparation for drawing the bonds according to their bond order (1, 2 or 3).
It currently draws every bond as if they're a single bond.

MolReader reads it from the MOL format (it's the 3rd value on each of the bond
lines), and this gets passed to the BondDto object to go to the client.