Make the ajax call a POST request
This way I can put the smiles string into the body
of the request and not have to worry about special
characters in the url

Note that Web API expects the POST body without a
key name, ('=value' not key=value). To achieve this
I pass in a JSON object with an empty property name
to the jquery ajax call.